.claude/skills/code-prompt-coach/ppt-generator/requirements.txt
python-pptx>=1.0.0
Pillow>=3.3.2
lxml>=3.1.0
